+++ packages/cli/create-medusa-app/src/utils/project-creator/index.ts
@@ -1,0 +1,4 @@
+export * from "./creator.js"
+export * from "./medusa-plugin-creator.js"
+export * from "./medusa-project-creator.js"
+export * from "./project-creator-factory.js"
